Migrate hmac encoding to an object
