+++ kmp-library/src/commonMain/kotlin/com/novoda/aws/v4/signer/hash/Hmac256Encoder.kt
@@ -1,0 +1,5 @@
+package com.novoda.aws.v4.signer.hash
+
+expect object Hmac256Encoder {
+    fun encode(key: ByteArray, value: String) : ByteArray
+}
