+++ kmp-library/src/iosMain/kotlin/com/novoda/aws/v4/signer/hash/Hmac256Encoder.kt
@@ -1,0 +1,7 @@
+package com.novoda.aws.v4.signer.hash
+
+actual object Hmac256Encoder {
+    actual fun encode(key: ByteArray, value: String): ByteArray {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
