--- kmp-library/src/iosMain/kotlin/com/novoda/aws/v4/signer/hash/toHmacSha256.kt
@@ -1,5 +1,0 @@
-package com.novoda.aws.v4.signer.hash
-
-actual fun String.toHmacSha256(key: ByteArray): ByteArray {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-}
